improve pipelines
git-svn-id: https://josm.openstreetmap.de/osmsvn/applications/editors/josm/plugins@34886 b9d5c4c9-76e1-0310-9c85-f3177eceb1e4

+++ seachart/jrender/src/jrender/Jrender.java
@@ -167,19 +167,19 @@ public final class Jrender {
             clean(12, 0, 0);
         }
         tile(zoom, 1, 0, 0);
-        if (send.size() > 0) {
-            PrintWriter writer = new PrintWriter(srcdir + zoom + "-" + xtile + "-" + ytile + ".send", "UTF-8");
-            for (String str : send) {
-                writer.println(str);
+        if ((send.size() > 0) || (deletes.size() > 0)) {
+        	PrintWriter writer = new PrintWriter(srcdir + zoom + "-" + xtile + "-" + ytile + ".send", "UTF-8");
+        	if (send.size() > 0) {
+        		for (String str : send) {
+        			writer.println(str);
             }
-            writer.close();
-        }
-        if (deletes.size() > 0) {
-            PrintWriter writer = new PrintWriter(srcdir + zoom + "-" + xtile + "-" + ytile + ".delete", "UTF-8");
-            for (String del : deletes.keySet()) {
-                writer.println("rm " + del);
+          }
+          if (deletes.size() > 0) {
+          	for (String del : deletes.keySet()) {
+          		writer.println("rm " + del);
             }
-            writer.close();
+          }
+          writer.close();
         }
         System.exit(0);
     }
